Feat: Making shift ids unique upon creation
Notes: This is not foolproof. It is dependent on the correctness of both my sorting algorithm, but also may break if the order of shifts are changed

Implimentaion: Pray that they dont change the order of events of with multiple of the same role. Uses the order of appearance of each role and assigns it a number
--Feat: Removed excess, moved to utils. from CMOEvent